// src/main/resources/angular/src/app/api/model/permission.ts
import { ScheduleCityConf } from "./scheduleCityConf";
import { SchedulePermission } from "./schedulePermission";

/**
 * Dum API
 * This is the DUM API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: kfiertek@nexusgeographics.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface Permission { 
    /**
     * ID permision
     */
    permissionId: number;
    name: string;
    quantity: string;
    expirationDate: string;
    periodicityId: number;
    cityId: number;
    /**
     * Tipus de conveni
     */
    agreementType: number;
    timeInMinutes: number;
    // to use in bo
    schedule?: SchedulePermission;
}